packages/python/infermap/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "infermap"
version = "0.3.2"
description = "Inference-driven schema mapping engine"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Ben Severn", email = "ben@bensevern.dev" }]
keywords = ["schema", "mapping", "etl", "data-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "polars>=1.0",
    "rapidfuzz>=3.0",
    "scipy>=1.11",
    "pyyaml>=6.0",
    "typer>=0.9",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]
mysql = ["mysql-connector-python>=8.0"]
duckdb = ["duckdb>=0.9"]
excel = ["openpyxl>=3.1"]
llm = ["openai>=1.0", "anthropic>=0.20"]
all = ["infermap[postgres,mysql,duckdb,excel,llm]"]
mcp = ["mcp>=1.0"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[project.scripts]
infermap = "infermap.cli:app"

[project.urls]
Homepage = "https://github.com/benzsevern/infermap"
Repository = "https://github.com/benzsevern/infermap"
Issues = "https://github.com/benzsevern/infermap/issues"
Documentation = "https://benzsevern.github.io/infermap/"
Changelog = "https://github.com/benzsevern/infermap/blob/main/CHANGELOG.md"
Author = "https://bensevern.dev"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
